+++ pkg/storegateway/bucket_store_metrics.go
@@ -2,151 +2,194 @@
 // Provenance-includes-location: https://github.com/cortexproject/cortex/blob/master/pkg/storegateway/bucket_store_metrics.go
 // Provenance-includes-license: Apache-2.0
 // Provenance-includes-copyright: The Cortex Authors.
+// Provenance-includes-location: https://github.com/thanos-io/thanos/blob/main/pkg/store/bucket.go
+// Provenance-includes-license: Apache-2.0
+// Provenance-includes-copyright: The Thanos Authors.
 
 package storegateway
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
 
 	"github.com/grafana/mimir/pkg/util"
 )
 
-// BucketStoreMetrics aggregates metrics exported by Thanos Bucket Store
-// and re-exports those aggregates as Mimir metrics.
+// BucketStoreMetrics holds all the metrics tracked by BucketStore. These metrics
+// MUST be monotonic (counter, summary, histogram) because a single metrics instance
+// can be passed to multiple BucketStore and metrics MUST be correct even after a
+// BucketStore is offloaded.
 type BucketStoreMetrics struct {
+	blockLoads            prometheus.Counter
+	blockLoadFailures     prometheus.Counter
+	blockDrops            prometheus.Counter
+	blockDropFailures     prometheus.Counter
+	seriesDataTouched     *prometheus.SummaryVec
+	seriesDataFetched     *prometheus.SummaryVec
+	seriesDataSizeTouched *prometheus.SummaryVec
+	seriesDataSizeFetched *prometheus.SummaryVec
+	seriesBlocksQueried   prometheus.Summary
+	seriesGetAllDuration  prometheus.Histogram
+	seriesMergeDuration   prometheus.Histogram
+	resultSeriesCount     prometheus.Summary
+	chunkSizeBytes        prometheus.Histogram
+	queriesDropped        *prometheus.CounterVec
+	seriesRefetches       prometheus.Counter
+
+	cachedPostingsCompressions           *prometheus.CounterVec
+	cachedPostingsCompressionErrors      *prometheus.CounterVec
+	cachedPostingsCompressionTimeSeconds *prometheus.CounterVec
+	cachedPostingsOriginalSizeBytes      prometheus.Counter
+	cachedPostingsCompressedSizeBytes    prometheus.Counter
+
+	seriesHashCacheRequests prometheus.Counter
+	seriesHashCacheHits     prometheus.Counter
+
+	seriesFetchDuration   prometheus.Histogram
+	postingsFetchDuration prometheus.Histogram
+}
+
+func NewBucketStoreMetrics(reg prometheus.Registerer) *BucketStoreMetrics {
+	var m BucketStoreMetrics
+
+	m.blockLoads = promauto.With(reg).NewCounter(prometheus.CounterOpts{
+		Name: "cortex_bucket_store_block_loads_total",
+		Help: "Total number of remote block loading attempts.",
+	})
+	m.blockLoadFailures = promauto.With(reg).NewCounter(prometheus.CounterOpts{
+		Name: "cortex_bucket_store_block_load_failures_total",
+		Help: "Total number of failed remote block loading attempts.",
+	})
+	m.blockDrops = promauto.With(reg).NewCounter(prometheus.CounterOpts{
+		Name: "cortex_bucket_store_block_drops_total",
+		Help: "Total number of local blocks that were dropped.",
+	})
+	m.blockDropFailures = promauto.With(reg).NewCounter(prometheus.CounterOpts{
+		Name: "cortex_bucket_store_block_drop_failures_total",
+		Help: "Total number of local blocks that failed to be dropped.",
+	})
+
+	m.seriesDataTouched = promauto.With(reg).NewSummaryVec(prometheus.SummaryOpts{
+		Name: "cortex_bucket_store_series_data_touched",
+		Help: "How many items of a data type in a block were touched for a single series request.",
+	}, []string{"data_type"})
+	m.seriesDataFetched = promauto.With(reg).NewSummaryVec(prometheus.SummaryOpts{
+		Name: "cortex_bucket_store_series_data_fetched",
+		Help: "How many items of a data type in a block were fetched for a single series request.",
+	}, []string{"data_type"})
+
+	m.seriesDataSizeTouched = promauto.With(reg).NewSummaryVec(prometheus.SummaryOpts{
+		Name: "cortex_bucket_store_series_data_size_touched_bytes",
+		Help: "Size of all items of a data type in a block were touched for a single series request.",
+	}, []string{"data_type"})
+	m.seriesDataSizeFetched = promauto.With(reg).NewSummaryVec(prometheus.SummaryOpts{
+		Name: "cortex_bucket_store_series_data_size_fetched_bytes",
+		Help: "Size of all items of a data type in a block were fetched for a single series request.",
+	}, []string{"data_type"})
+
+	m.seriesBlocksQueried = promauto.With(reg).NewSummary(prometheus.SummaryOpts{
+		Name: "cortex_bucket_store_series_blocks_queried",
+		Help: "Number of blocks in a bucket store that were touched to satisfy a query.",
+	})
+	m.seriesGetAllDuration = promauto.With(reg).NewHistogram(prometheus.HistogramOpts{
+		Name:    "cortex_bucket_store_series_get_all_duration_seconds",
+		Help:    "Time it takes until all per-block prepares and loads for a query are finished.",
+		Buckets: []float64{0.001, 0.01, 0.1, 0.3, 0.6, 1, 3, 6, 9, 20, 30, 60, 90, 120},
+	})
+	m.seriesMergeDuration = promauto.With(reg).NewHistogram(prometheus.HistogramOpts{
+		Name:    "cortex_bucket_store_series_merge_duration_seconds",
+		Help:    "Time it takes to merge sub-results from all queried blocks into a single result.",
+		Buckets: []float64{0.001, 0.01, 0.1, 0.3, 0.6, 1, 3, 6, 9, 20, 30, 60, 90, 120},
+	})
+	m.seriesRefetches = promauto.With(reg).NewCounter(prometheus.CounterOpts{
+		Name: "cortex_bucket_store_series_refetches_total",
+		Help: "Total number of cases where the built-in max series size was not enough to fetch series from index, resulting in refetch.",
+	})
+	m.resultSeriesCount = promauto.With(reg).NewSummary(prometheus.SummaryOpts{
+		Name: "cortex_bucket_store_series_result_series",
+		Help: "Number of series observed in the final result of a query.",
+	})
+	m.queriesDropped = promauto.With(reg).NewCounterVec(prometheus.CounterOpts{
+		Name: "cortex_bucket_store_queries_dropped_total",
+		Help: "Number of queries that were dropped due to the max chunks per query limit.",
+	}, []string{"reason"})
+
+	m.cachedPostingsCompressions = promauto.With(reg).NewCounterVec(prometheus.CounterOpts{
+		Name: "cortex_bucket_store_cached_postings_compressions_total",
+		Help: "Number of postings compressions and decompressions when storing to index cache.", // TODO also decompressions?
+	}, []string{"op"})
+	m.cachedPostingsCompressions.WithLabelValues(labelEncode)
+	m.cachedPostingsCompressions.WithLabelValues(labelDecode)
+
+	m.cachedPostingsCompressionErrors = promauto.With(reg).NewCounterVec(prometheus.CounterOpts{
+		Name: "cortex_bucket_store_cached_postings_compression_errors_total",
+		Help: "Number of postings compression and decompression errors.",
+	}, []string{"op"})
+	m.cachedPostingsCompressionErrors.WithLabelValues(labelEncode)
+	m.cachedPostingsCompressionErrors.WithLabelValues(labelDecode)
+
+	m.cachedPostingsCompressionTimeSeconds = promauto.With(reg).NewCounterVec(prometheus.CounterOpts{
+		Name: "cortex_bucket_store_cached_postings_compression_time_seconds",
+		Help: "Time spent compressing and decompressing postings when storing to / reading from postings cache.",
+	}, []string{"op"})
+	m.cachedPostingsCompressionTimeSeconds.WithLabelValues(labelEncode)
+	m.cachedPostingsCompressionTimeSeconds.WithLabelValues(labelDecode)
+
+	m.cachedPostingsOriginalSizeBytes = promauto.With(reg).NewCounter(prometheus.CounterOpts{
+		Name: "cortex_bucket_store_cached_postings_original_size_bytes_total",
+		Help: "Original size of postings stored into cache.",
+	})
+	m.cachedPostingsCompressedSizeBytes = promauto.With(reg).NewCounter(prometheus.CounterOpts{
+		Name: "cortex_bucket_store_cached_postings_compressed_size_bytes_total",
+		Help: "Compressed size of postings stored into cache.",
+	})
+
+	m.seriesFetchDuration = promauto.With(reg).NewHistogram(prometheus.HistogramOpts{
+		Name:    "cortex_bucket_store_cached_series_fetch_duration_seconds",
+		Help:    "Time it takes to fetch series to respond a request sent to store-gateway. It includes both the time to fetch it from cache and from storage in case of cache misses.",
+		Buckets: []float64{0.001, 0.01, 0.1, 0.3, 0.6, 1, 3, 6, 9, 20, 30, 60, 90, 120},
+	})
+	m.postingsFetchDuration = promauto.With(reg).NewHistogram(prometheus.HistogramOpts{
+		Name:    "cortex_bucket_store_cached_postings_fetch_duration_seconds",
+		Help:    "Time it takes to fetch postings to respond a request sent to store-gateway. It includes both the time to fetch it from cache and from storage in case of cache misses.",
+		Buckets: []float64{0.001, 0.01, 0.1, 0.3, 0.6, 1, 3, 6, 9, 20, 30, 60, 90, 120},
+	})
+
+	m.seriesHashCacheRequests = promauto.With(reg).NewCounter(prometheus.CounterOpts{
+		Name: "cortex_bucket_store_series_hash_cache_requests_total",
+		Help: "Total number of fetch attempts to the in-memory series hash cache.",
+	})
+	m.seriesHashCacheHits = promauto.With(reg).NewCounter(prometheus.CounterOpts{
+		Name: "cortex_bucket_store_series_hash_cache_hits_total",
+		Help: "Total number of fetch hits to the in-memory series hash cache.",
+	})
+
+	m.chunkSizeBytes = promauto.With(reg).NewHistogram(prometheus.HistogramOpts{
+		Name: "cortex_bucket_store_sent_chunk_size_bytes",
+		Help: "Size in bytes of the chunks for the single series, which is adequate to the gRPC message size sent to querier.",
+		Buckets: []float64{
+			32, 256, 512, 1024, 32 * 1024, 256 * 1024, 512 * 1024, 1024 * 1024, 32 * 1024 * 1024, 256 * 1024 * 1024, 512 * 1024 * 1024,
+		},
+	})
+
+	return &m
+}
+
+// IndexHeaderMetrics aggregates metrics exported by Thanos index-header lazy loader
+// and re-exports those aggregates as Mimir metrics.
+type IndexHeaderMetrics struct {
 	regs *util.UserRegistries
-
-	// exported metrics, gathered from Thanos BucketStore
-	blockLoads            *prometheus.Desc
-	blockLoadFailures     *prometheus.Desc
-	blockDrops            *prometheus.Desc
-	blockDropFailures     *prometheus.Desc
-	blocksLoaded          *prometheus.Desc
-	seriesDataTouched     *prometheus.Desc
-	seriesDataFetched     *prometheus.Desc
-	seriesDataSizeTouched *prometheus.Desc
-	seriesDataSizeFetched *prometheus.Desc
-	seriesBlocksQueried   *prometheus.Desc
-	seriesGetAllDuration  *prometheus.Desc
-	seriesMergeDuration   *prometheus.Desc
-	seriesRefetches       *prometheus.Desc
-	resultSeriesCount     *prometheus.Desc
-	queriesDropped        *prometheus.Desc
-
-	cachedPostingsCompressions           *prometheus.Desc
-	cachedPostingsCompressionErrors      *prometheus.Desc
-	cachedPostingsCompressionTimeSeconds *prometheus.Desc
-	cachedPostingsOriginalSizeBytes      *prometheus.Desc
-	cachedPostingsCompressedSizeBytes    *prometheus.Desc
-
-	seriesFetchDuration   *prometheus.Desc
-	postingsFetchDuration *prometheus.Desc
 
 	indexHeaderLazyLoadCount         *prometheus.Desc
 	indexHeaderLazyLoadFailedCount   *prometheus.Desc
 	indexHeaderLazyUnloadCount       *prometheus.Desc
 	indexHeaderLazyUnloadFailedCount *prometheus.Desc
 	indexHeaderLazyLoadDuration      *prometheus.Desc
-
-	seriesHashCacheRequests *prometheus.Desc
-	seriesHashCacheHits     *prometheus.Desc
 }
 
-func NewBucketStoreMetrics() *BucketStoreMetrics {
-	return &BucketStoreMetrics{
+func NewIndexHeaderMetrics() *IndexHeaderMetrics {
+	return &IndexHeaderMetrics{
 		regs: util.NewUserRegistries(),
-
-		blockLoads: prometheus.NewDesc(
-			"cortex_bucket_store_block_loads_total",
-			"Total number of remote block loading attempts.",
-			nil, nil),
-		blockLoadFailures: prometheus.NewDesc(
-			"cortex_bucket_store_block_load_failures_total",
-			"Total number of failed remote block loading attempts.",
-			nil, nil),
-		blockDrops: prometheus.NewDesc(
-			"cortex_bucket_store_block_drops_total",
-			"Total number of local blocks that were dropped.",
-			nil, nil),
-		blockDropFailures: prometheus.NewDesc(
-			"cortex_bucket_store_block_drop_failures_total",
-			"Total number of local blocks that failed to be dropped.",
-			nil, nil),
-		blocksLoaded: prometheus.NewDesc(
-			"cortex_bucket_store_blocks_loaded",
-			"Number of currently loaded blocks.",
-			nil, nil),
-		seriesDataTouched: prometheus.NewDesc(
-			"cortex_bucket_store_series_data_touched",
-			"How many items of a data type in a block were touched for a single series request.",
-			[]string{"data_type"}, nil),
-		seriesDataFetched: prometheus.NewDesc(
-			"cortex_bucket_store_series_data_fetched",
-			"How many items of a data type in a block were fetched for a single series request.",
-			[]string{"data_type"}, nil),
-		seriesDataSizeTouched: prometheus.NewDesc(
-			"cortex_bucket_store_series_data_size_touched_bytes",
-			"Size of all items of a data type in a block were touched for a single series request.",
-			[]string{"data_type"}, nil),
-		seriesDataSizeFetched: prometheus.NewDesc(
-			"cortex_bucket_store_series_data_size_fetched_bytes",
-			"Size of all items of a data type in a block were fetched for a single series request.",
-			[]string{"data_type"}, nil),
-		seriesBlocksQueried: prometheus.NewDesc(
-			"cortex_bucket_store_series_blocks_queried",
-			"Number of blocks in a bucket store that were touched to satisfy a query.",
-			nil, nil),
-
-		seriesGetAllDuration: prometheus.NewDesc(
-			"cortex_bucket_store_series_get_all_duration_seconds",
-			"Time it takes until all per-block prepares and preloads for a query are finished.",
-			nil, nil),
-		seriesMergeDuration: prometheus.NewDesc(
-			"cortex_bucket_store_series_merge_duration_seconds",
-			"Time it takes to merge sub-results from all queried blocks into a single result.",
-			nil, nil),
-		seriesRefetches: prometheus.NewDesc(
-			"cortex_bucket_store_series_refetches_total",
-			"Total number of cases where the built-in max series size was not enough to fetch series from index, resulting in refetch.",
-			nil, nil),
-		resultSeriesCount: prometheus.NewDesc(
-			"cortex_bucket_store_series_result_series",
-			"Number of series observed in the final result of a query.",
-			nil, nil),
-		queriesDropped: prometheus.NewDesc(
-			"cortex_bucket_store_queries_dropped_total",
-			"Number of queries that were dropped due to the max chunks per query limit.",
-			nil, nil),
-
-		cachedPostingsCompressions: prometheus.NewDesc(
-			"cortex_bucket_store_cached_postings_compressions_total",
-			"Number of postings compressions and decompressions when storing to index cache.",
-			[]string{"op"}, nil),
-		cachedPostingsCompressionErrors: prometheus.NewDesc(
-			"cortex_bucket_store_cached_postings_compression_errors_total",
-			"Number of postings compression and decompression errors.",
-			[]string{"op"}, nil),
-		cachedPostingsCompressionTimeSeconds: prometheus.NewDesc(
-			"cortex_bucket_store_cached_postings_compression_time_seconds",
-			"Time spent compressing and decompressing postings when storing to / reading from postings cache.",
-			[]string{"op"}, nil),
-		cachedPostingsOriginalSizeBytes: prometheus.NewDesc(
-			"cortex_bucket_store_cached_postings_original_size_bytes_total",
-			"Original size of postings stored into cache.",
-			nil, nil),
-		cachedPostingsCompressedSizeBytes: prometheus.NewDesc(
-			"cortex_bucket_store_cached_postings_compressed_size_bytes_total",
-			"Compressed size of postings stored into cache.",
-			nil, nil),
-
-		seriesFetchDuration: prometheus.NewDesc(
-			"cortex_bucket_store_cached_series_fetch_duration_seconds",
-			"Time it takes to fetch series to respond a request sent to store-gateway. It includes both the time to fetch it from cache and from storage in case of cache misses.",
-			nil, nil),
-		postingsFetchDuration: prometheus.NewDesc(
-			"cortex_bucket_store_cached_postings_fetch_duration_seconds",
-			"Time it takes to fetch postings to respond a request sent to store-gateway. It includes both the time to fetch it from cache and from storage in case of cache misses.",
-			nil, nil),
 
 		indexHeaderLazyLoadCount: prometheus.NewDesc(
 			"cortex_bucket_store_indexheader_lazy_load_total",
@@ -168,99 +211,31 @@ func NewBucketStoreMetrics() *BucketStoreMetrics {
 			"cortex_bucket_store_indexheader_lazy_load_duration_seconds",
 			"Duration of the index-header lazy loading in seconds.",
 			nil, nil),
-
-		seriesHashCacheRequests: prometheus.NewDesc(
-			"cortex_bucket_store_series_hash_cache_requests_total",
-			"Total number of fetch attempts to the in-memory series hash cache.",
-			nil, nil),
-		seriesHashCacheHits: prometheus.NewDesc(
-			"cortex_bucket_store_series_hash_cache_hits_total",
-			"Total number of fetch hits to the in-memory series hash cache.",
-			nil, nil),
 	}
 }
 
-func (m *BucketStoreMetrics) AddUserRegistry(user string, reg *prometheus.Registry) {
+func (m *IndexHeaderMetrics) AddUserRegistry(user string, reg *prometheus.Registry) {
 	m.regs.AddUserRegistry(user, reg)
 }
 
-func (m *BucketStoreMetrics) RemoveUserRegistry(user string) {
+func (m *IndexHeaderMetrics) RemoveUserRegistry(user string) {
 	m.regs.RemoveUserRegistry(user, false)
 }
 
-func (m *BucketStoreMetrics) Describe(out chan<- *prometheus.Desc) {
-	out <- m.blockLoads
-	out <- m.blockLoadFailures
-	out <- m.blockDrops
-	out <- m.blockDropFailures
-	out <- m.blocksLoaded
-	out <- m.seriesDataTouched
-	out <- m.seriesDataFetched
-	out <- m.seriesDataSizeTouched
-	out <- m.seriesDataSizeFetched
-	out <- m.seriesBlocksQueried
-	out <- m.seriesGetAllDuration
-	out <- m.seriesMergeDuration
-	out <- m.seriesRefetches
-	out <- m.resultSeriesCount
-	out <- m.queriesDropped
-
-	out <- m.cachedPostingsCompressions
-	out <- m.cachedPostingsCompressionErrors
-	out <- m.cachedPostingsCompressionTimeSeconds
-	out <- m.cachedPostingsOriginalSizeBytes
-	out <- m.cachedPostingsCompressedSizeBytes
-
-	out <- m.seriesFetchDuration
-	out <- m.postingsFetchDuration
-
+func (m *IndexHeaderMetrics) Describe(out chan<- *prometheus.Desc) {
 	out <- m.indexHeaderLazyLoadCount
 	out <- m.indexHeaderLazyLoadFailedCount
 	out <- m.indexHeaderLazyUnloadCount
 	out <- m.indexHeaderLazyUnloadFailedCount
 	out <- m.indexHeaderLazyLoadDuration
-
-	out <- m.seriesHashCacheRequests
-	out <- m.seriesHashCacheHits
 }
 
-func (m *BucketStoreMetrics) Collect(out chan<- prometheus.Metric) {
+func (m *IndexHeaderMetrics) Collect(out chan<- prometheus.Metric) {
 	data := m.regs.BuildMetricFamiliesPerUser()
-
-	data.SendSumOfCounters(out, m.blockLoads, "thanos_bucket_store_block_loads_total")
-	data.SendSumOfCounters(out, m.blockLoadFailures, "thanos_bucket_store_block_load_failures_total")
-	data.SendSumOfCounters(out, m.blockDrops, "thanos_bucket_store_block_drops_total")
-	data.SendSumOfCounters(out, m.blockDropFailures, "thanos_bucket_store_block_drop_failures_total")
-
-	data.SendSumOfGauges(out, m.blocksLoaded, "thanos_bucket_store_blocks_loaded")
-
-	data.SendSumOfSummariesWithLabels(out, m.seriesDataTouched, "thanos_bucket_store_series_data_touched", "data_type")
-	data.SendSumOfSummariesWithLabels(out, m.seriesDataFetched, "thanos_bucket_store_series_data_fetched", "data_type")
-	data.SendSumOfSummariesWithLabels(out, m.seriesDataSizeTouched, "thanos_bucket_store_series_data_size_touched_bytes", "data_type")
-	data.SendSumOfSummariesWithLabels(out, m.seriesDataSizeFetched, "thanos_bucket_store_series_data_size_fetched_bytes", "data_type")
-	data.SendSumOfSummariesWithLabels(out, m.seriesBlocksQueried, "thanos_bucket_store_series_blocks_queried")
-
-	data.SendSumOfHistograms(out, m.seriesGetAllDuration, "thanos_bucket_store_series_get_all_duration_seconds")
-	data.SendSumOfHistograms(out, m.seriesMergeDuration, "thanos_bucket_store_series_merge_duration_seconds")
-	data.SendSumOfCounters(out, m.seriesRefetches, "thanos_bucket_store_series_refetches_total")
-	data.SendSumOfSummaries(out, m.resultSeriesCount, "thanos_bucket_store_series_result_series")
-	data.SendSumOfCounters(out, m.queriesDropped, "thanos_bucket_store_queries_dropped_total")
-
-	data.SendSumOfCountersWithLabels(out, m.cachedPostingsCompressions, "thanos_bucket_store_cached_postings_compressions_total", "op")
-	data.SendSumOfCountersWithLabels(out, m.cachedPostingsCompressionErrors, "thanos_bucket_store_cached_postings_compression_errors_total", "op")
-	data.SendSumOfCountersWithLabels(out, m.cachedPostingsCompressionTimeSeconds, "thanos_bucket_store_cached_postings_compression_time_seconds_total", "op")
-	data.SendSumOfCountersWithLabels(out, m.cachedPostingsOriginalSizeBytes, "thanos_bucket_store_cached_postings_original_size_bytes_total")
-	data.SendSumOfCountersWithLabels(out, m.cachedPostingsCompressedSizeBytes, "thanos_bucket_store_cached_postings_compressed_size_bytes_total")
-
-	data.SendSumOfHistograms(out, m.seriesFetchDuration, "thanos_bucket_store_cached_series_fetch_duration_seconds")
-	data.SendSumOfHistograms(out, m.postingsFetchDuration, "thanos_bucket_store_cached_postings_fetch_duration_seconds")
 
 	data.SendSumOfCounters(out, m.indexHeaderLazyLoadCount, "thanos_bucket_store_indexheader_lazy_load_total")
 	data.SendSumOfCounters(out, m.indexHeaderLazyLoadFailedCount, "thanos_bucket_store_indexheader_lazy_load_failed_total")
 	data.SendSumOfCounters(out, m.indexHeaderLazyUnloadCount, "thanos_bucket_store_indexheader_lazy_unload_total")
 	data.SendSumOfCounters(out, m.indexHeaderLazyUnloadFailedCount, "thanos_bucket_store_indexheader_lazy_unload_failed_total")
 	data.SendSumOfHistograms(out, m.indexHeaderLazyLoadDuration, "thanos_bucket_store_indexheader_lazy_load_duration_seconds")
-
-	data.SendSumOfCounters(out, m.seriesHashCacheRequests, "thanos_bucket_store_series_hash_cache_requests_total")
-	data.SendSumOfCounters(out, m.seriesHashCacheHits, "thanos_bucket_store_series_hash_cache_hits_total")
 }